requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
aiosqlite==0.19.0
pytz==2023.3
python-dateutil==2.8.2
pathlib==1.0.1

